Refactor ssgan example
