+++ examples/ssgan/utils.py
